cpufreq: Check all online CPUs on input event
Previously, only the first CPU in an SMP system was bumped to the
maximum clock when receiving an input event. This worked fine on
uniprocessor systems or scenarios were other CPUs tended to be
powered down. In other situations, it is better to bump up the
clocks of all CPUs that are online. This will not affect the
situations mentioned above, but it will help others.

Signed-off-by: Brian Steuer <bsteuer@codeaurora.org>

Conflicts:
	drivers/cpufreq/cpufreq_ondemand.c

Change-Id: Iac76eeb3306dd6b968b98cb61c092b2711b79c48